Long 타입 테스트. long 값이 작으면(8bit: -126 ~ 127) ==으로 동일 판별 가능. 크면(8bit 넘는 값) ==으로 불가. longValue() 해야 함. 만약 userId가 8bit를 넘어가게 되면, == 비교로 checkOwner()를 수행하면 안됨. -> 디버깅이 안됨. -> if(user.getId().longValue() != userId.longValue())으로 수정. => 테스트는 정말 꼼꼼히 해야 함. 내가 아는 것만 해서는 버그가 생길 확률이 높음. => 여러 번의 체크 또는 QA 정도의 집착이 필요한 것 같음.

+++ bank/src/main/java/shop/mtcoding/bank/domain/account/Account.java
@@ -67,7 +67,7 @@ public class Account {
     public void checkOwner(Long userId) {
         // String testUsername = user.getUsername(); // lazy 로딩되어야 함.
         // System.out.println("테스트 : " + testUsername);
-        if(user.getId() != userId) { // lazy 로딩이어도 id를 조회할 때는 select 쿼리가 날라가지 않음.
+        if(user.getId().longValue() != userId.longValue()) { // lazy 로딩이어도 id를 조회할 때는 select 쿼리가 날라가지 않음.
             throw new CustomApiException("계좌 소유자가 아닙니다.");
         }
     }
